+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 112.4.0 (Jun 1, 2023)
+
+### Minor
+
+- GlobalEventsHandlerProvider: new provider to share global event handlers to consuming components ([#2959](https://github.com/pinterest/gestalt/pull/2959)) - [Preview link](https://deploy-preview-2959--gestalt.netlify.app)
+
+## 112.3.9 (Jun 1, 2023)
+
+### Patch
+
+- Table: Convert examples to use Sandpack ([#2969](https://github.com/pinterest/gestalt/pull/2969)) - [Preview link](https://deploy-preview-2969--gestalt.netlify.app)
+
+## 112.3.8 (Jun 1, 2023)
+
+### Patch
+
+- Docs: Added Android specs for Avatar, Button, Checkbox, IconButton, Searchfield, Sheet, Switch, Tabs, Textfield ([#2972](https://github.com/pinterest/gestalt/pull/2972)) - [Preview link](https://deploy-preview-2972--gestalt.netlify.app)
+
 ## 112.3.7 (May 30, 2023)
 
 ### Patch